Work around `python3` and `spk dev` tracing issues
At runtime, Python 3 stat's a lot of .py files, and compares the timestamp
against the .pyc files.  If the .pyc file is newer, it'll use the .pyc file and
ignore the .py file.

`spk dev` doesn't add files that get stat()'d but not open()'d to
sandstorm-files.list, because that would probably pull in a huge number of
unrelated paths.

However, the existence of the .py files is important - if they are absent, then
Python 3 will refuse to load the related .pyc files.  But with `spk dev`s
tracing today, they will not get included in the package.

It seems simpler to just include all of /usr/lib/python3.4 (and a couple other
paths relevant to Python packages), which also means that changes will cause
less unnecessary churn.  The only issue with that is that it pulls in some
large files, namely, 55MB of static libs for python from the Debian python3-dev
package.

So to avoid bloating the package with this data, we remove the python3-dev
package after we're done installing python dependencies from pip, and then
everything is great - we have a small-ish (15MB) .spk that works in dev mode as
well as packed.